Add public getters to theme data and brightness
The actual theme data and brightness members are still private and can only be modified using setBrightness() and setThemeData().

But making the current values public enables using them for UI toggles like checkboxes, which need to access the current value.
Otherwise, we are forced to introduce further state to keep track of the current brightness.